Move TextNodeNamespace to correct packge
